Tried to undo a commit github.
Unfortunately this stemmed from trying to undo a commit to github due to having the wrong email associated with it. A rebase was done and forgot to do a git stash. This resulted in loss of today's work. The database needed to be dropped and started over to get everything synced up again.